refactor: clean up indentity structure #58
