+++ src/Account/appsettings.json
@@ -78,6 +78,7 @@
     }
   },
   "ConnectionStrings": {
-    "Account": "Server=localhost;Port=5490;Database=profile_db;User Id=postgres;Password=postgres;Include Error Detail=true"
+    "Account": "Server=localhost;Port=5490;Database=profile_db;User Id=postgres;Password=postgres;Include Error Detail=true",
+    "Identity": "Server=localhost;Port=5490;Database=identity_db;User Id=postgres;Password=postgres;Include Error Detail=true"
   }
 }
