+++ src/Common/Core.Framework/Core.Framework.csproj
@@ -4,6 +4,8 @@
     </PropertyGroup>
     <ItemGroup>
         <FrameworkReference Include="Microsoft.AspNetCore.App" />
+        <PackageReference Include="NWebsec.AspNetCore.Core" />
+        <PackageReference Include="NWebsec.AspNetCore.Middleware" />
     </ItemGroup>
     <ItemGroup>
       <ProjectReference Include="..\Core.Logging\Core.Logging.csproj" />
