+++ src/Identity/appsettings.json
@@ -58,5 +58,8 @@
   },
   "app": {
     "name": "Identity Service"
+  },
+  "ConnectionStrings": {
+    "Identity": "Server=localhost;Port=5490;Database=identity_db;User Id=postgres;Password=postgres;Include Error Detail=true"
   }
 }
